+++ apps/web/src/app/docs/components/menu/examples/usage.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { Button, Menu } from "@jamsrui/react";
 
 export const MenuUsage = () => {
@@ -8,13 +7,13 @@ export const MenuUsage = () => {
         <Button>Open Me</Button>
       </Menu.Trigger>
       <Menu.Content>
-        <Menu.Item>Undo</Menu.Item>
-        <Menu.Item>Info</Menu.Item>
-        <Menu.Item>Search</Menu.Item>
-        <Menu.Item>Redo</Menu.Item>
-        <Menu.Item>Cut</Menu.Item>
-        <Menu.Item>Edit</Menu.Item>
-        <Menu.Item>Delete</Menu.Item>
+        <Menu.Item textValue="Undo">Undo</Menu.Item>
+        <Menu.Item textValue="Info">Info</Menu.Item>
+        <Menu.Item textValue="Search">Search</Menu.Item>
+        <Menu.Item textValue="Redo">Redo</Menu.Item>
+        <Menu.Item textValue="Cut">Cut</Menu.Item>
+        <Menu.Item textValue="Edit">Edit</Menu.Item>
+        <Menu.Item textValue="Delete">Delete</Menu.Item>
       </Menu.Content>
     </Menu>
   );