+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsrui/accordion
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/composite@0.0.7
+  - @jamsrui/core@0.0.12
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes