+++ packages/components/alert/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsrui/alert
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/text@0.0.14
+  - @jamsrui/core@0.0.12
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/icons@0.0.12
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes