+++ packages/components/linear-progress/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsrui/linear-progress
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/core@0.0.12
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes