+++ packages/components/link/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsrui/link
 
+## 0.0.15
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/text@0.0.14
+  - @jamsrui/core@0.0.12
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.14
 
 ### Patch Changes