+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsrui/switch
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/slot@0.0.13
+  - @jamsrui/context@0.0.3
+  - @jamsrui/core@0.0.12
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes