+++ packages/components/textfield/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsrui/textfield
 
+## 0.0.3
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/context@0.0.3
+  - @jamsrui/hooks@0.0.14
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.2
 
 ### Patch Changes