+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsrui/hooks
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+- Updated dependencies
+  - @jamsrui/utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes