+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsrui/icons
 
+## 0.0.12
+
+### Patch Changes
+
+- fix; exports
+
 ## 0.0.11
 
 ### Patch Changes