+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsrui/utils
 
+## 0.0.14
+
+### Patch Changes
+
+- fix; exports
+
 ## 0.0.13
 
 ### Patch Changes